Fix Topbar and userToggle
